requirements.txt
fastapi
uvicorn
python-multipart
pydantic
torch
torchvision
torchaudio
timm
librosa
pillow
matplotlib
numpy
soundfile
